refactor(client): split lib.rs into client module structure
Split the large lib.rs file (1532 lines) into a well-organized client
module with three files:
- src/client/rest.rs (420 lines) - Low-level REST HTTP client
- src/client/websocket.rs (147 lines) - WebSocket event stream
- src/client/mod.rs (1086 lines) - Main Webex client

Key changes:
- Reduced lib.rs from 1532 to 54 lines (thin orchestrator)
- Added AuthorizationType enum for flexible auth (None, Bearer, Basic)
- Added api_post_form_urlencoded for OAuth device grant flow
- Fixed GET requests to use query parameters instead of body
- Fixed empty response handling for 204 No Content
- Made client module public for auth.rs access
- Updated auth.rs imports to use new module structure
- All 37 tests passing
